Move snapshot_steps tests into their own file
Split test_snapshot_steps_matches_full_history_final_state,
test_snapshot_steps_reduces_peak_gpu_memory, and their shared subprocess
helper out of tests/test_simulation.py into tests/test_snapshot_steps.py,
keeping test_simulation.py focused on the pre-existing Simulation surface.
Imports small_simulation_parameters from tests.test_simulation to avoid
duplicating the helper.
